+++ src/RcppExports.cpp
@@ -5,18 +5,6 @@
 
 using namespace Rcpp;
 
-// fast_table
-IntegerVector fast_table(IntegerVector x, int size);
-RcppExport SEXP _rlas_fast_table(SEXP xSEXP, SEXP sizeSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< IntegerVector >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type size(sizeSEXP);
-    rcpp_result_gen = Rcpp::wrap(fast_table(x, size));
-    return rcpp_result_gen;
-END_RCPP
-}
 // fast_countequal
 int fast_countequal(IntegerVector x, int t);
 RcppExport SEXP _rlas_fast_countequal(SEXP xSEXP, SEXP tSEXP) {
@@ -113,7 +101,6 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_rlas_fast_table", (DL_FUNC) &_rlas_fast_table, 2},
     {"_rlas_fast_countequal", (DL_FUNC) &_rlas_fast_countequal, 2},
     {"_rlas_fast_countbelow", (DL_FUNC) &_rlas_fast_countbelow, 2},
     {"_rlas_fast_countover", (DL_FUNC) &_rlas_fast_countover, 2},