+++ src/fast.cpp
@@ -2,20 +2,6 @@
 using namespace Rcpp;
 
 // [[Rcpp::export]]
-IntegerVector fast_table(IntegerVector x, int size = 5)
-{
-  IntegerVector tbl(size);
-
-  for (IntegerVector::iterator it = x.begin(), end = x.end() ; it != end ; ++it)
-  {
-    if (*it <= size && *it > 0)
-      tbl(*it-1)++;
-  }
-
-  return tbl;
-}
-
-// [[Rcpp::export]]
 int fast_countequal(IntegerVector x, int t)
 {
   return std::count(x.begin(), x.end(), t);